Upgrades backend

+++ backend/controllers/product.controller.js
@@ -1,4 +1,5 @@
 import Product from "../model/product.model.js";
+import mongoose from "mongoose";
 
 //GET
 export const getProducts = async (req, res) => {
@@ -56,11 +57,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
     
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({succes:false, message:"Invalid Product Id"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({succes: true, message: "Product deleted"});
     } catch (error) {
         console.log("error in deleting product:", error.message);
-        res.status(404).json({succes: false, message: "Product not found"});
+        res.status(500).json({succes: false, message: "Server Error"});
     }
 };
